자리바꾸기 전 코드 정리

+++ solve/0729/java_hw_08_2_sol/src/com/ssafy/hw/step2/UserManagerImpl.java
@@ -98,8 +98,7 @@ public class UserManagerImpl implements IUserManager {
 		for (int i = 0; i < userList.size(); i++) {
 			sum += userList.get(i).getAge();
 		}
-
-		return sum / userList.size();
+		return (double)sum / userList.size();
 
 	}
